Aligned with current implementation

+++ enclave-Alice/eapp/eapp_crt.h
@@ -1,7 +1,7 @@
 #ifndef _EAPP_CRT_H_
 #define _EAPP_CRT_H_
 
-int store_crt(unsigned char *crt, int crt_len);
-int read_crt(unsigned char *crt, int *crt_len);
+int store_crt(unsigned char *crt, size_t crt_len);
+int read_crt(unsigned char *crt, size_t *crt_len);
 
 #endif
